move parser implementation in separate file
