+++ JSON/Makefile
@@ -10,9 +10,9 @@ include $(POCO_BASE)/build/rules/global
 
 INCLUDE += -I $(POCO_BASE)/JSON/include/Poco/JSON
 
-objects = Array Object Parser Handler Stringifier \
-	ParseHandler PrintHandler Query JSONException \
-	Template TemplateCache
+objects = Array Object Parser ParserImpl Handler \
+	Stringifier ParseHandler PrintHandler Query \
+	JSONException Template TemplateCache
 
 target         = PocoJSON
 target_version = $(LIBVERSION)
